# packages/dashboard/openapi.yaml
openapi: 3.1.0
info:
  title: Aurora Dashboard API
  version: 0.3.0
  description: >
    Admin-plane API for aurora.local. Same-origin, cookie-authenticated
    (session cookie set by POST /auth/login). All paths below are served
    under the `/api` prefix by the Spring Boot backend; the Vite dev
    server proxies `/api` to :8090.


    This spec is hand-authored from the typed frontend `src/api/*`
    modules and is the source of truth the MSW mock layer answers to.
    Field casing is deliberately mixed: most bodies are camelCase, but a
    handful of endpoints emit snake_case verbatim from the backend
    (onboarding draft/plan/launch, services, containers logs, security
    dismissals, audit). Casing here matches the wire, not a convention.
servers:
  - url: /api
    description: Same-origin, behind Caddy / Vite proxy.
...
security:
  - sessionCookie: []
paths:
...
  /onboarding:
    get:
      tags: [onboarding]
      summary: Full onboarding draft (hydration)
      security: []
      responses:
        '200':
          description: Draft.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/OnboardingDraft' }
    patch:
      tags: [onboarding]
      summary: Partial update of any draft field
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/OnboardingPatch' }
      responses:
        '200':
          description: Updated draft.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/OnboardingDraft' }
        '409':
          description: No admin yet, or wizard already committed.
...
  /onboarding/admin:
    post:
      tags: [onboarding]
      summary: One-shot bootstrap of the initial admin
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [username, password]
              properties:
                username: { type: string }
                password: { type: string }
      responses:
        '204': { description: Admin created. }
        '409': { description: Admin already exists. }
  /onboarding/plan:
    get:
      tags: [onboarding]
      summary: Install plan (optionally scoped to a speculative enabled set)
      security: []
      parameters:
        - name: enabled
          in: query
          required: false
          description: CSV of package names to preview instead of on-disk state.
          schema: { type: string }
      responses:
        '200':
          description: Plan (snake_case wire shape).
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PlanWire' }
...
  /onboarding/launch/{id}:
    get:
      tags: [onboarding]
      summary: Snapshot of a launch job
      parameters:
        - $ref: '#/components/parameters/JobId'
      responses:
        '200':
          description: Launch status.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/LaunchStatus' }
        '404': { description: Unknown job. }
  /onboarding/launch/{id}/stream:
    get:
      tags: [onboarding]
      summary: SSE — live launch log + terminal state
      description: >
        Server-Sent Events. Emits `log` events (one line per message),
        `ping` keepalives, and a terminal `done` event carrying the final
        LaunchStatus.
      parameters:
        - $ref: '#/components/parameters/JobId'
      responses:
        '200':
          description: Event stream.
          content:
            text/event-stream:
              schema: { type: string }
...
  /onboarding/domain:
    post:
      tags: [onboarding]
      summary: Set the domain during the wizard
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                domain: { type: string }
      responses:
        '204': { description: Saved. }
        '400': { description: The domain is not usable. }
  /onboarding/packages:
    post:
      tags: [onboarding]
      summary: Set the enabled packages during the wizard
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                packages:
                  type: array
                  items: { type: string }
      responses:
        '204': { description: Saved. }
  /onboarding/sso:
    post:
      tags: [onboarding]
      summary: Set the single sign-on choice during the wizard
      description: Phase D. Decides whether Authelia fronts the apps that support it.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
      responses:
        '204': { description: Saved. }

  # ────────────────────────────── updates ─────────────────────────────
  # Read-only. Applying an update goes through
  # POST /packages/{name}/upgrade, which already exists and already
  # returns a job id — there is deliberately no second update verb.
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: AURORA_SESSION
  parameters:
...
    JobId:
      name: id
      in: path
      required: true
      schema: { type: string }
...
  schemas:
...
    OnboardingStepId:
      type: string
      enum: [welcome, admin, domain, packages, secrets, dns, tls, review, done]
    DnsMode:
      type: string
      enum: [adguard, router, mdns]
    OnboardingDraft:
      type: object
      required: [complete, bootstrap_mode, step, admin_username, domain, enabled_packages, dns_mode]
      properties:
        complete: { type: boolean }
        bootstrap_mode: { type: boolean }
        step: { $ref: '#/components/schemas/OnboardingStepId' }
        admin_username: { type: [string, 'null'] }
        domain: { type: [string, 'null'] }
        enabled_packages:
          type: array
          items: { type: string }
        dns_mode:
          oneOf:
            - { $ref: '#/components/schemas/DnsMode' }
            - { type: 'null' }
...
    OnboardingPatch:
      type: object
      properties:
        domain: { type: string }
        enabled_packages: { type: array, items: { type: string } }
        dns_mode: { $ref: '#/components/schemas/DnsMode' }
        step: { $ref: '#/components/schemas/OnboardingStepId' }
...
    PlanWire:
      type: object
      required: [packages_to_enable, packages_to_disable, vhosts, ports, warnings]
      properties:
        packages_to_enable: { type: array, items: { type: string } }
        packages_to_disable: { type: array, items: { type: string } }
        vhosts: { type: array, items: { type: string } }
        ports: { type: array, items: { type: integer } }
        warnings: { type: array, items: { type: string } }
...
    LaunchStatus:
      type: object
      required: [id, state, packages, started_at, finished_at, exit_code, failure_reason, tail]
      properties:
        id: { type: string }
        state: { type: string, enum: [running, success, failed] }
        packages: { type: array, items: { type: string } }
        started_at: { type: string, format: date-time }
        finished_at: { type: [string, 'null'], format: date-time }
        exit_code: { type: [integer, 'null'] }
        failure_reason: { type: [string, 'null'] }
        tail: { type: array, items: { type: string } }
...